+++ src/apphosting/secrets/index.ts
@@ -8,6 +8,16 @@ import { FIREBASE_MANAGED } from "../../gcp/secretManager";
 import { isFunctionsManaged } from "../../gcp/secretManager";
 import * as utils from "../../utils";
 import * as prompt from "../../prompt";
+import { basename } from "path";
+import {
+  APPHOSTING_BASE_YAML_FILE,
+  APPHOSTING_LOCAL_YAML_FILE,
+  APPHOSTING_YAML_FILE_REGEX,
+  discoverConfigsAtBackendRoot,
+  loadConfigForEnvironment,
+} from "../config";
+import { promptForAppHostingYaml } from "../utils";
+import { AppHostingYamlConfig, Secret } from "../yaml";
 
 /** Interface for holding the service account pair for a given Backend. */
 export interface ServiceAccounts {
@@ -163,4 +173,106 @@ export async function upsertSecret(
   // TODO: consider whether we should prompt a user who has an unmanaged secret to enroll in version control.
   // This may not be a great idea until version control is actually implemented.
   return false;
+}
+
+/**
+ * Fetches secrets from Google Secret Manager and returns their values in plain text.
+ */
+export async function fetchSecrets(
+  projectId: string,
+  secrets: Secret[],
+): Promise<Map<string, string>> {
+  let secretsKeyValuePairs: Map<string, string>;
+
+  try {
+    const secretPromises: Promise<[string, string]>[] = secrets.map(async (secretConfig) => {
+      const [name, version] = getSecretNameParts(secretConfig.secret!);
+
+      const value = await gcsm.accessSecretVersion(projectId, name, version);
+      return [secretConfig.variable, value] as [string, string];
+    });
+
+    const secretEntries = await Promise.all(secretPromises);
+    secretsKeyValuePairs = new Map(secretEntries);
+  } catch (e: any) {
+    throw new FirebaseError(`Error exporting secrets`, {
+      original: e,
+    });
+  }
+
+  return secretsKeyValuePairs;
+}
+
+/**
+ * Returns the appropriate App Hosting YAML configuration for exporting secrets.
+ *
+ * @returns The final merged config
+ */
+export async function loadConfigToExport(
+  cwd: string,
+  userGivenConfigFile?: string,
+): Promise<AppHostingYamlConfig> {
+  if (userGivenConfigFile && !APPHOSTING_YAML_FILE_REGEX.test(userGivenConfigFile)) {
+    throw new FirebaseError(
+      "Invalid apphosting yaml config file provided. File must be in format: 'apphosting.yaml' or 'apphosting.<environment>.yaml'",
+    );
+  }
+
+  const allConfigs = discoverConfigs(cwd);
+  let userGivenConfigFilePath: string;
+  if (userGivenConfigFile) {
+    if (!allConfigs.has(userGivenConfigFile)) {
+      throw new FirebaseError(
+        `The provided app hosting config file "${userGivenConfigFile}" does not exist`,
+      );
+    }
+
+    userGivenConfigFilePath = allConfigs.get(userGivenConfigFile)!;
+  } else {
+    userGivenConfigFilePath = await promptForAppHostingYaml(
+      allConfigs,
+      "Which environment would you like to export secrets from Secret Manager for?",
+    );
+  }
+
+  if (userGivenConfigFile === APPHOSTING_BASE_YAML_FILE) {
+    return AppHostingYamlConfig.loadFromFile(allConfigs.get(APPHOSTING_BASE_YAML_FILE)!);
+  }
+
+  const baseFilePath = allConfigs.get(APPHOSTING_BASE_YAML_FILE)!;
+  return await loadConfigForEnvironment(userGivenConfigFilePath, baseFilePath);
+}
+
+/**
+ * Gets all apphosting yaml configs excluding apphosting.local.yaml and returns
+ * a map in the format {"apphosting.staging.yaml" => "/cwd/apphosting.staging.yaml"}.
+ */
+function discoverConfigs(cwd: string): Map<string, string> {
+  const appHostingConfigPaths = discoverConfigsAtBackendRoot(cwd).filter(
+    (path) => !path.endsWith(APPHOSTING_LOCAL_YAML_FILE),
+  );
+  if (appHostingConfigPaths.length === 0) {
+    throw new FirebaseError("No apphosting.*.yaml configs found");
+  }
+
+  const fileNameToPathMap: Map<string, string> = new Map();
+  for (const path of appHostingConfigPaths) {
+    const fileName = basename(path);
+    fileNameToPathMap.set(fileName, path);
+  }
+
+  return fileNameToPathMap;
+}
+/**
+ * secret expected to be in format "myApiKeySecret@5",
+ * "projects/test-project/secrets/secretID", or
+ * "projects/test-project/secrets/secretID/versions/5"
+ */
+export function getSecretNameParts(secret: string): [string, string] {
+  let [name, version] = secret.split("@");
+  if (!version) {
+    version = "latest";
+  }
+
+  return [name, version];
 }