+++ src/emulator/dataconnectEmulator.ts
@@ -1,5 +1,6 @@
 import * as childProcess from "child_process";
 import { EventEmitter } from "events";
+import * as clc from "colorette";
 
 import { dataConnectLocalConnString } from "../api";
 import { Constants } from "./constants";
@@ -65,13 +66,13 @@ export class DataConnectEmulator implements EmulatorInstance {
         if (Constants.isDemoProject(this.args.projectId)) {
           this.logger.logLabeled(
             "WARN",
-            "Data Connect",
+            "dataconnect",
             "Detected a 'demo-' project, but vector embeddings require a real project. Operations that use vector_embed will fail.",
           );
         } else {
           this.logger.logLabeled(
             "WARN",
-            "Data Connect",
+            "dataconnect",
             "Operations that use vector_embed will make calls to production Vertex AI",
           );
         }
@@ -80,9 +81,6 @@ export class DataConnectEmulator implements EmulatorInstance {
       this.logger.log("DEBUG", `'fdc build' failed with error: ${err.message}`);
     }
 
-    const info = await load(this.args.projectId, this.args.config, this.args.configDir);
-    const dbId = info.dataConnectYaml.schema.datasource.postgresql?.database || "postgres";
-    const serviceId = info.dataConnectYaml.serviceId;
     await start(Emulators.DATACONNECT, {
       auto_download: this.args.auto_download,
       listen: listenSpecsToString(this.args.listen),
@@ -92,14 +90,17 @@ export class DataConnectEmulator implements EmulatorInstance {
     });
     this.usingExistingEmulator = false;
     if (this.args.autoconnectToPostgres) {
+      const info = await load(this.args.projectId, this.args.config, this.args.configDir);
+      const dbId = info.dataConnectYaml.schema.datasource.postgresql?.database || "postgres";
+      const serviceId = info.dataConnectYaml.serviceId;
       const pgPort = this.args.postgresListen?.[0].port;
       const pgHost = this.args.postgresListen?.[0].address;
       let connStr = dataConnectLocalConnString();
-      if (dataConnectLocalConnString()) {
+      if (connStr) {
         this.logger.logLabeled(
           "INFO",
-          "Data Connect",
-          `FIREBASE_DATACONNECT_POSTGRESQL_STRING is set to ${dataConnectLocalConnString()} - using that instead of starting a new database`,
+          "dataconnect",
+          `FIREBASE_DATACONNECT_POSTGRESQL_STRING is set to ${clc.bold(connStr)} - using that instead of starting a new database`,
         );
       } else if (pgHost && pgPort) {
         const dataDirectory = this.args.config.get("emulators.dataconnect.dataDirectory");
@@ -114,7 +115,7 @@ export class DataConnectEmulator implements EmulatorInstance {
             console;
             this.logger.logLabeled(
               "ERROR",
-              "Data Connect",
+              "dataconnect",
               `Postgres threw an unexpected error, shutting down the Data Connect emulator: ${err}`,
             );
           }
@@ -122,8 +123,8 @@ export class DataConnectEmulator implements EmulatorInstance {
         });
         this.logger.logLabeled(
           "INFO",
-          "Data Connect",
-          `Started up Postgres server, listening on ${server.address()?.toString()}`,
+          "dataconnect",
+          `Started up Postgres server, listening on ${JSON.stringify(server.address())}`,
         );
       }
       await this.connectToPostgres(new URL(connStr), dbId, serviceId);
@@ -137,7 +138,7 @@ export class DataConnectEmulator implements EmulatorInstance {
     if (!emuInfo) {
       this.logger.logLabeled(
         "ERROR",
-        "Data Connect",
+        "dataconnect",
         "Could not connect to Data Connect emulator. Check dataconnect-debug.log for more details.",
       );
       return Promise.reject();
@@ -149,7 +150,7 @@ export class DataConnectEmulator implements EmulatorInstance {
     if (this.usingExistingEmulator) {
       this.logger.logLabeled(
         "INFO",
-        "Data Connect",
+        "dataconnect",
         "Skipping cleanup of Data Connect emulator, as it was not started by this process.",
       );
       return;