+++ src/emulator/downloadableEmulators.ts
@@ -59,20 +59,20 @@ const EMULATOR_UPDATE_DETAILS: { [s in DownloadableEmulators]: EmulatorUpdateDet
   dataconnect:
     process.platform === "darwin"
       ? {
-          version: "1.5.1",
-          expectedSize: 25289472,
-          expectedChecksum: "92c425072db66c7e2cfa40b703ed807b",
+          version: "1.7.0",
+          expectedSize: 25350912,
+          expectedChecksum: "1479411a52b689a990179c7f674362b6",
         }
       : process.platform === "win32"
         ? {
-            version: "1.5.1",
-            expectedSize: 25720320,
-            expectedChecksum: "2a5c654770233b740980d5f98f24be73",
+            version: "1.7.0",
+            expectedSize: 25783808,
+            expectedChecksum: "fef62bf25e816aa4ba250e80050f0125",
           }
         : {
-            version: "1.5.1",
-            expectedSize: 25202840,
-            expectedChecksum: "f95156cbcac237268791638ea0eb10e7",
+            version: "1.7.0",
+            expectedSize: 25272472,
+            expectedChecksum: "24d873457787546ca6f2470fe9ec3edd",
           },
 };
 
@@ -281,8 +281,8 @@ const Commands: { [s in DownloadableEmulators]: DownloadableEmulatorCommand } = 
     shell: true,
   },
   ui: {
-    binary: "node",
-    args: [getExecPath(Emulators.UI)],
+    binary: "",
+    args: [],
     optionalArgs: [],
     joinArgs: false,
     shell: false,